Update package export signature
